haizzz

+++ src/data_pipeline/preprocessing/video_pose_extractor.py
@@ -3,45 +3,41 @@ import cv2
 import pandas as pd
 import os
 from tqdm import tqdm
-from pose_detect import *
+from src.data_pipeline.preprocessing.pose_detect import *
 
 
 def resize_videos_and_save(dataset_path, output_path, target_size=(640, 640)):
+    if os.path.exists(output_path):
+        print(f"Output path {output_path} already exists. Skipping resizing.")
+        return
+
     os.makedirs(output_path, exist_ok=True)
 
-    for split in ['Train', 'Test']:
-        split_path = os.path.join(dataset_path, split)
-        if not os.path.exists(split_path):
-            raise ValueError(f"Not found {split} folder in {dataset_path}")
+    for class_folder in os.listdir(dataset_path):
+        class_path = os.path.join(dataset_path, class_folder)
 
-        output_split_path = os.path.join(output_path, split)
-        os.makedirs(output_split_path, exist_ok=True)
+        # Check if it is a directory
+        if os.path.isdir(class_path):
+            # Create class folder in output path
+            output_class_path = os.path.join(output_path, class_folder)
+            os.makedirs(output_class_path, exist_ok=True)
 
-        for class_folder in os.listdir(split_path):
-            class_path = os.path.join(split_path, class_folder)
+            for video in os.listdir(class_path):
+                video_path = os.path.join(class_path, video)
+                video_ext = os.path.splitext(video)[1]
 
-            # Check if it is a directory
-            if os.path.isdir(class_path):
-                # Create class folder in output path
-                output_class_path = os.path.join(output_split_path, class_folder)
-                os.makedirs(output_class_path, exist_ok=True)
-
-                for video in os.listdir(class_path):
-                    video_path = os.path.join(class_path, video)
-                    video_ext = os.path.splitext(video)[1]
-
-                    if os.path.isfile(video_path) and video_ext in ['.mp4', '.avi', '.mov']:
-                        try:
-                            clip = VideoFileClip(video_path)
-                            resized_clip = clip.resize(target_size)
-                            new_video_name = f"resized_{video}"
-                            new_video_path = os.path.join(output_class_path, new_video_name)
-                            resized_clip.write_videofile(new_video_path, codec="libx264")
-                            clip.close()
-                            resized_clip.close()
-                            print(f"Handled video {video_path}")
-                        except Exception as e:
-                            raise ValueError(f"Error when handling video {video_path}: {e}")
+                if os.path.isfile(video_path) and video_ext in ['.mp4', '.avi', '.mov']:
+                    try:
+                        clip = VideoFileClip(video_path)
+                        resized_clip = clip.resize(target_size)
+                        new_video_name = f"resized_{video}"
+                        new_video_path = os.path.join(output_class_path, new_video_name)
+                        resized_clip.write_videofile(new_video_path, codec="libx264")
+                        clip.close()
+                        resized_clip.close()
+                        print(f"Handled video {video_path}")
+                    except Exception as e:
+                        raise ValueError(f"Error when handling video {video_path}: {e}")
 
 
 def detect_pose_video(video_path,
@@ -66,8 +62,8 @@ def detect_pose_video(video_path,
         timestamp = frame_num / fps
         # Convert the image to RGB for MediaPipe
         image = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-        # Process the image with MediaPipe Holistic model (without face detection)
-        row_data = get_pose_image(image, frame_num, timestamp)
+        # Process the image with MediaPipe Holistic models (without face detection)
+        row_data = get_pose_image(image)
         # Append the row data to the list
         landmark_data.append(row_data)
         # Move to the next frame
@@ -88,24 +84,41 @@ def detect_pose_video(video_path,
     return df
 
 
-def process_videos(source_root: str, destination_root: str):
+def process_videos(resize_dataset_video_path, destination_path):
     X, y = [], []
-    for dir_path, dir_names, filenames in os.walk(source_root):
-        relative_dir: str = os.path.relpath(dir_path, source_root)
-        dest_dir = os.path.join(destination_root, relative_dir)
 
-        for filename in tqdm(filenames, desc=f"Processing {relative_dir}"):
+    if not os.path.exists(resize_dataset_video_path):
+        raise ValueError(f"Path {resize_dataset_video_path} does not exist")
+
+    print(f'Processing videos from {os.path.abspath(resize_dataset_video_path)}')
+
+    for class_folder in os.listdir(resize_dataset_video_path):
+        dir_path = os.path.join(resize_dataset_video_path, class_folder)
+
+        # Check if it is a directory
+        if not os.path.isdir(os.path.join(destination_path, class_folder)):
+            os.makedirs(os.path.join(destination_path, class_folder), exist_ok=True)
+
+        filenames = os.listdir(dir_path)
+        for filename in tqdm(filenames, desc=f"Processing {class_folder}"):
             if filename.lower().endswith(('.mp4', '.avi', '.mov', '.mkv')):
                 video_path = os.path.join(dir_path, filename)
 
                 csv_filename = os.path.splitext(filename)[0] + '.csv'  # Change extension to .csv
-                csv_path = os.path.join(dest_dir, csv_filename)  # Create the path for the CSV file
+                csv_path = os.path.join(destination_path, class_folder, csv_filename)  # Create the path for the CSV file
 
+                # Check if the CSV file already exists
+                if os.path.exists(csv_path):
+                    df = pd.read_csv(csv_path)
+                    for column in df.columns:
+                        df[column] = df[column].apply(lambda x: np.array(x[1:-1].split(',')).astype(np.float32))
+                    X.append(df)
+                    y.append(class_folder)
+                    continue
                 try:
                     df = detect_pose_video(video_path)  # Process the video
                     X.append(df)  # Append the DataFrame to the list
-                    y.append(relative_dir)  # Append the label to the list
-
+                    y.append(class_folder)  # Append the class to the list
                     df.to_csv(csv_path, index=False)
                 except Exception as e:
                     raise ValueError(f"Error when processing video {video_path}: {e}")
